More rogue run mechanics

+++ spellsource-server/src/main/java/com/hiddenswitch/framework/impl/RogueManager.java
@@ -13,6 +13,7 @@ import com.hiddenswitch.framework.schema.spellsource.tables.records.RogueRunReco
 import com.hiddenswitch.spellsource.rpc.Spellsource.CardTypeMessage.CardType;
 import io.vertx.core.Future;
 import io.vertx.sqlclient.Row;
+import net.demilich.metastone.game.cards.Attribute;
 import net.demilich.metastone.game.cards.Card;
 import net.demilich.metastone.game.cards.CardArrayList;
 import net.demilich.metastone.game.entities.heroes.HeroClass;
@@ -24,6 +25,7 @@ import org.jooq.ResultQuery;
 import org.jooq.UpdateSetFirstStep;
 import org.jooq.UpdateSetMoreStep;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Function;
 
@@ -40,18 +42,23 @@ public class RogueManager {
 		cardCatalogue.invalidateAllAndRefresh();
 	}
 
-	public static Future<Long> startRogueRun(String heroClass, long seed, String userId) {
-		var rogueRun = await(Environment.callRoutine(Routines.startRogueRun(heroClass, seed)).withUserId(userId).execute(RowMappers.getRogueRunMapper()));
+	public static Future<Long> startRogueRun(String heroClass, Long seed, String userId) {
+		var rogueRun =
+			await(Environment.callRoutine(Routines.startRogueRun(heroClass, seed)).withUserId(userId).execute(RowMappers.getRogueRunMapper()));
+
+		seed = rogueRun.getSeed();
 
 		var random = new XORShiftRandom(seed);
 
-		var deckCards = getCardChoices(random, rogueRun.getHeroClass(), RogueChoiceType.STANDARD, 10).toArray(String[]::new);
+		var deckCards = new ArrayList<>(getCardChoices(random, rogueRun.getHeroClass(), RogueChoiceType.STANDARD, 15));
+		deckCards.add("level_1");
 
-		await(Environment.callRoutine(Routines.setCardsInDeck(rogueRun.getDeck(), deckCards)).execute(RowMappers.getCardsInDeckMapper()));
+		await(Environment.callRoutine(Routines.setCardsInDeck(rogueRun.getDeck(), deckCards.toArray(String[]::new))).execute(RowMappers.getCardsInDeckMapper()));
 
 		await(updateOpponentDeck(rogueRun, random));
 
-		await(updateRogueRun(rogueRun.getId(), r -> r.set(ROGUE_RUN.STATE, RogueRunState.PRE_MATCH).set(ROGUE_RUN.SEED_STATE, random.getState())));
+		await(updateRogueRun(rogueRun.getId(),
+			r -> r.set(ROGUE_RUN.STATE, RogueRunState.PRE_MATCH).set(ROGUE_RUN.SEED_STATE, random.getState())));
 
 		return Future.succeededFuture(rogueRun.getId());
 	}
@@ -87,10 +94,12 @@ public class RogueManager {
 			var userId = rogueRun.getPlayer();
 			var deckId = rogueRun.getDeck();
 
-			var deckCards = await(Environment.callRoutine(Routines.getCardsInDeck(rogueRun.getDeck())).execute(row -> row.getString(0)));
+			var deckCards =
+				await(Environment.callRoutine(Routines.getCardsInDeck(rogueRun.getDeck())).execute(row -> row.getString(0)));
 			var deck = new CardArrayList(deckCards.stream().map(cardCatalogue::getCardById).toList());
 
-			var context = new RogueChoiceGameContext(cardCatalogue, rogueRun.getHeroClass(), deck, userId, deckId, rogueRun.getSeed());
+			var context = new RogueChoiceGameContext(cardCatalogue, rogueRun.getHeroClass(), deck, userId, deckId,
+				rogueRun.getSeed());
 			context.init(); // TODO add info for rogue run id ?
 			var player = context.getPlayer1();
 
@@ -109,7 +118,8 @@ public class RogueManager {
 				}
 
 				if (card.getDesc().getRogueInfo() != null && card.getDesc().getRogueInfo().getChosenSpell() != null) {
-					context.getLogic().castSpell(PLAYER_1, card.getDesc().getRogueInfo().getChosenSpell(), card.getReference(), null, TargetSelection.NONE, false, null);
+					context.getLogic().castSpell(PLAYER_1, card.getDesc().getRogueInfo().getChosenSpell(), card.getReference(),
+						null, TargetSelection.NONE, false, null);
 				}
 
 				context.getLogic().fireGameEvent(new RogueChoiceEvent(context, player, card));
@@ -117,14 +127,16 @@ public class RogueManager {
 
 			seedState = context.getLogic().getRandom().getState();
 
-			await(Environment.callRoutine(Routines.setCardsInDeck(deckId, context.getPlayer1().getDeck().stream().map(Card::getCardId).toArray(String[]::new))).execute(RowMappers.getCardsInDeckMapper()));
+			await(Environment.callRoutine(Routines.setCardsInDeck(deckId,
+				context.getPlayer1().getDeck().stream().map(Card::getCardId).toArray(String[]::new))).execute(RowMappers.getCardsInDeckMapper()));
 		}
 
 		if (!choice.getRepopulate() || choices.size() >= choice.getCanPick()) {
 			await(Environment.withDslContext(dsl -> dsl.deleteFrom(ROGUE_CHOICE).where(ROGUE_CHOICE.ID.eq(choiceId))));
 		} else {
 			var random = new XORShiftRandom(seedState);
-			var newCards = getCardChoices(random, rogueRun.getHeroClass(), choice.getType(), choice.getCanPick() - choices.size());
+			var newCards = getCardChoices(random, rogueRun.getHeroClass(), choice.getType(),
+				choice.getCanPick() - choices.size());
 			seedState = random.getState();
 
 			var newChoices = choice.getCards();
@@ -132,13 +144,16 @@ public class RogueManager {
 				newChoices[choices.get(i)] = newCards.get(i);
 			}
 
-			await(updateRogueChoice(choiceId, c -> c.set(ROGUE_CHOICE.CARDS, newChoices).set(ROGUE_CHOICE.CAN_PICK, choice.getCanPick() - choices.size())));
+			await(updateRogueChoice(choiceId, c -> c.set(ROGUE_CHOICE.CARDS, newChoices).set(ROGUE_CHOICE.CAN_PICK,
+				choice.getCanPick() - choices.size())));
 		}
 
-		var newChoice = await(Environment.callRoutine(Routines.currentRogueChoice(rogueRun.getId())).execute(RowMappers.getRogueChoiceMapper()));
+		var newChoice =
+			await(Environment.callRoutine(Routines.currentRogueChoice(rogueRun.getId())).execute(RowMappers.getRogueChoiceMapper()));
 
 		var finalSeedState = seedState;
-		await(updateRogueRun(rogueRun.getId(), r -> r.set(ROGUE_RUN.STATE, newChoice == null ? RogueRunState.PRE_MATCH : RogueRunState.CHOICE).set(ROGUE_RUN.SEED_STATE, finalSeedState)));
+		await(updateRogueRun(rogueRun.getId(), r -> r.set(ROGUE_RUN.STATE, newChoice == null ? RogueRunState.PRE_MATCH :
+			RogueRunState.CHOICE).set(ROGUE_RUN.SEED_STATE, finalSeedState)));
 
 		return Future.succeededFuture(rogueRun.getId());
 	}
@@ -168,10 +183,12 @@ public class RogueManager {
 
 		var random = new XORShiftRandom(rogueRun.getSeedState());
 
-		var newChoices = getCardChoices(random, rogueRun.getHeroClass(), choice.getType(), choice.getCards().length).toArray(String[]::new);
+		var newChoices =
+			getCardChoices(random, rogueRun.getHeroClass(), choice.getType(), choice.getCards().length).toArray(String[]::new);
 
 		await(updateRogueChoice(choiceId, c -> c.set(ROGUE_CHOICE.CARDS, newChoices)));
-		await(updateRogueRun(rogueRun.getId(), r -> r.set(ROGUE_RUN.GOLD, rogueRun.getGold() - cost).set(ROGUE_RUN.SEED_STATE, random.getState())));
+		await(updateRogueRun(rogueRun.getId(),
+			r -> r.set(ROGUE_RUN.GOLD, rogueRun.getGold() - cost).set(ROGUE_RUN.SEED_STATE, random.getState())));
 
 		return Future.succeededFuture(rogueRun.getId());
 	}
@@ -187,9 +204,14 @@ public class RogueManager {
 	}
 
 	public static Future<Long> trashCard(Long rogueId, String cardId) {
+		if (cardId.startsWith("level_")) {
+			return Future.failedFuture("Can't trash that card");
+		}
+
 		var rogueRun = await(getRogueRun(rogueId));
 
-		var count = await(Environment.withDslContext(dsl -> dsl.deleteFrom(CARDS_IN_DECK).where(CARDS_IN_DECK.DECK_ID.eq(rogueRun.getDeck()).and(CARDS_IN_DECK.CARD_ID.eq(cardId))).limit(1)));
+		var count =
+			await(Environment.withDslContext(dsl -> dsl.deleteFrom(CARDS_IN_DECK).where(CARDS_IN_DECK.DECK_ID.eq(rogueRun.getDeck()).and(CARDS_IN_DECK.CARD_ID.eq(cardId))).limit(1)));
 
 		if (count < 1) {
 			return Future.failedFuture("Card did not exist within deck to trash");
@@ -203,7 +225,8 @@ public class RogueManager {
 	}
 
 	public static Future<RogueRun> handleGameEnd(long gameId, String winnerUserId) {
-		var result = await(Environment.callRoutine(Routines.checkRogueGameEnd(gameId, winnerUserId)).execute(RowMappers.getRogueRunMapper()));
+		var result =
+			await(Environment.callRoutine(Routines.checkRogueGameEnd(gameId, winnerUserId)).execute(RowMappers.getRogueRunMapper()));
 
 		if (result == null) {
 			return Future.succeededFuture(null);
@@ -214,12 +237,17 @@ public class RogueManager {
 
 		await(addNewRogueChoice(r -> r.setType(RogueChoiceType.EQUIPMENT).setRogueRun(result.getId()).setIndex(0).setCards(equipment.toArray(String[]::new)).setCanPick(1)));
 
-		var cards = getCardChoices(random, result.getHeroClass(), RogueChoiceType.EQUIPMENT, 3);
+		var cards = getCardChoices(random, result.getHeroClass(), RogueChoiceType.STANDARD, 3);
 		await(addNewRogueChoice(r -> r.setType(RogueChoiceType.STANDARD).setRogueRun(result.getId()).setIndex(1).setCards(cards.toArray(String[]::new)).setCanPick(3).setRepopulate(true).setCanReroll(true)));
 
 		await(updateOpponentDeck(result, random));
 
-		await(updateRogueRun(result.getId(), r -> r.set(ROGUE_RUN.GOLD, result.getGold() + result.getBossesDefeated()).set(ROGUE_RUN.SEED_STATE, random.getState())));
+		await(updateRogueRun(result.getId(),
+			r -> r.set(ROGUE_RUN.GOLD, result.getGold() + result.getBossesDefeated()).set(ROGUE_RUN.SEED_STATE,
+				random.getState())));
+
+		await(Environment.withDslContext(dsl -> dsl.update(CARDS_IN_DECK).set(CARDS_IN_DECK.CARD_ID,
+			"level_" + (result.getBossesDefeated() + 1)).where(CARDS_IN_DECK.CARD_ID.eq("level_" + result.getBossesDefeated()))));
 
 		return Future.succeededFuture(result);
 	}
@@ -230,17 +258,23 @@ public class RogueManager {
 		var deck = premadeDecks.get(random.nextInt(premadeDecks.size()));
 
 		await(Environment.withDslContext(dsl -> dsl.update(DECKS).set(DECKS.HERO_CLASS, deck.getHeroClass()).where(DECKS.ID.eq(rogueRun.getOpponentDeck()))));
-		await(Environment.callRoutine(Routines.setCardsInDeck(rogueRun.getOpponentDeck(), deck.getCardIds().toArray(String[]::new))).execute(RowMappers.getCardsInDeckMapper()));
+		await(Environment.callRoutine(Routines.setCardsInDeck(rogueRun.getOpponentDeck(),
+			deck.getCardIds().toArray(String[]::new))).execute(RowMappers.getCardsInDeckMapper()));
+
+		await(Environment.withDslContext(dsl -> dsl.insertInto(CARDS_IN_DECK, CARDS_IN_DECK.DECK_ID,
+			CARDS_IN_DECK.CARD_ID).values(rogueRun.getOpponentDeck(), "level_" + (rogueRun.getBossesDefeated() + 1))));
 
 		return Future.succeededFuture();
 	}
 
 	public static Future<RogueRun> returningRogueRun(Function<DSLContext, ResultQuery<RogueRunRecord>> handler) {
-		return Environment.withExecutor(executor -> executor.findOneRow(handler).map(row -> row == null ? null : RowMappers.getRogueRunMapper().apply(row)));
+		return Environment.withExecutor(executor -> executor.findOneRow(handler).map(row -> row == null ? null :
+			RowMappers.getRogueRunMapper().apply(row)));
 	}
 
 	public static Future<RogueChoice> returningRogueChoice(Function<DSLContext, ResultQuery<RogueChoiceRecord>> handler) {
-		return Environment.withExecutor(executor -> executor.findOneRow(handler).map(row -> row == null ? null : rogueChoiceMapper().apply(row)));
+		return Environment.withExecutor(executor -> executor.findOneRow(handler).map(row -> row == null ? null :
+			rogueChoiceMapper().apply(row)));
 	}
 
 	public static Future<RogueRun> getRogueRun(long rogueId) {
@@ -255,21 +289,33 @@ public class RogueManager {
 		return returningRogueChoice(dsl -> dsl.selectFrom(ROGUE_CHOICE).where(ROGUE_CHOICE.ID.eq(choiceId)));
 	}
 
-	public static Future<RogueRun> updateRogueRun(long rogueId, Function<UpdateSetFirstStep<RogueRunRecord>, UpdateSetMoreStep<RogueRunRecord>> handler) {
+	public static Future<RogueRun> updateRogueRun(long rogueId, Function<UpdateSetFirstStep<RogueRunRecord>,
+		UpdateSetMoreStep<RogueRunRecord>> handler) {
 		return returningRogueRun(dsl -> handler.apply(dsl.update(ROGUE_RUN)).where(ROGUE_RUN.ID.eq(rogueId)).returning());
 	}
 
-	public static Future<RogueChoice> updateRogueChoice(long choiceId, Function<UpdateSetFirstStep<RogueChoiceRecord>, UpdateSetMoreStep<RogueChoiceRecord>> handler) {
+	public static Future<RogueChoice> updateRogueChoice(long choiceId, Function<UpdateSetFirstStep<RogueChoiceRecord>,
+		UpdateSetMoreStep<RogueChoiceRecord>> handler) {
 		return returningRogueChoice(dsl -> handler.apply(dsl.update(ROGUE_CHOICE)).where(ROGUE_CHOICE.ID.eq(choiceId)).returning());
 	}
 
 	public static Function<Row, RogueChoice> rogueChoiceMapper() {
-		return row -> row == null ? null : RowMappers.getRogueChoiceMapper().apply(row).setCards(row.getArrayOfStrings("cards"));
+		return row -> row == null ? null : RowMappers.getRogueChoiceMapper().apply(row).setCards(row.getArrayOfStrings(
+			"cards"));
 	}
 
-	public static List<String> getCardChoices(XORShiftRandom random, String heroClass, RogueChoiceType choiceType, int howMany) {
+	public static boolean filterCard(RogueChoiceType choiceType, Card card) {
+		return switch (choiceType) {
+			case STANDARD -> card.isCollectible();
+			case EQUIPMENT -> card.hasAttribute(Attribute.EQUIPMENT);
+		};
+	}
+
+	public static List<String> getCardChoices(XORShiftRandom random, String heroClass, RogueChoiceType choiceType,
+																						int howMany) {
 		var format = cardCatalogue.getFormat("Rogue");
-		var cards = cardCatalogue.query(format, card -> (card.hasHeroClass(heroClass) || card.hasHeroClass(HeroClass.ANY)) && card.isCollectible());
+		var cards = cardCatalogue.query(format,
+			card -> (card.hasHeroClass(heroClass) || card.hasHeroClass(HeroClass.ANY)) && filterCard(choiceType, card));
 
 		cards.shuffle(random);
